Combined output functions in a single function
Combined PPOutWriteSqlite3, PPOutWriteHDF5 and PPOutWriteCSV in PPOutput. Made according changes to all the relevant files.
